requirements.txt
fastapi
uvicorn 
pydantic

flask
requests
python-dotenv==1.0.0